refactor(frontend): extract invites modal from main.ts (#312)
Continues the conservative split: pull the invite-code mint/list/revoke
modal (admin-only, #50) into invites.ts (~194 lines).

The global Escape-key handler that physically sits in the invites section is
NOT invites-specific (it dismisses the topmost of every modal) — it stays in
main.ts and now imports closeInvitesModal from invites.ts. Same TDZ-safe
pattern otherwise: DOM re-queried locally; the showToast import is used only
inside functions; the 3 shared invites consts (used by the Escape guard /
applyAdminVisibility) stay in main.ts.

main.ts: 3831 → 3651. tsc + Biome clean, vite build green, frontend tests
pass (66).
